DDP-6591 | remove not used test, moved UserDto to dto directory
