+++ src/main/java/org/broadinstitute/dsm/db/dto/user/UserDto.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsm.db;
+package org.broadinstitute.dsm.db.dto.user;
 
 import java.util.Optional;
