+++ src/test/java/org/broadinstitute/dsm/FilterTest.java
@@ -1,7 +1,7 @@
 package org.broadinstitute.dsm;
 
 import com.google.gson.Gson;
-import org.broadinstitute.dsm.db.UserDto;
+import org.broadinstitute.dsm.db.dto.user.UserDto;
 import org.broadinstitute.dsm.db.dao.ddp.instance.DDPInstanceDao;
 import org.broadinstitute.dsm.db.dao.ddp.participant.ParticipantDataDao;
 import org.broadinstitute.dsm.db.dao.user.UserDao;
