+++ src/test/java/org/broadinstitute/dsm/pubsub/UpdateWorkflowStatusTest.java
@@ -2,7 +2,7 @@ package org.broadinstitute.dsm.pubsub;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
-import org.broadinstitute.dsm.db.UserDto;
+import org.broadinstitute.dsm.db.dto.user.UserDto;
 import org.broadinstitute.dsm.db.dao.ddp.instance.DDPInstanceDao;
 import org.broadinstitute.dsm.db.dao.fieldsettings.FieldSettingsDao;
 import org.broadinstitute.dsm.db.dao.ddp.participant.ParticipantDataDao;
