+++ src/test/java/org/broadinstitute/dsm/route/familymember/AddFamilyMemberRouteTest.java
@@ -9,14 +9,12 @@ import java.util.NoSuchElementException;
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
 import org.broadinstitute.ddp.handlers.util.Result;
-import org.broadinstitute.dsm.db.UserDto;
+import org.broadinstitute.dsm.db.dto.user.UserDto;
 import org.broadinstitute.dsm.db.dao.ddp.instance.DDPInstanceDao;
 import org.broadinstitute.dsm.db.dao.ddp.participant.ParticipantDataDao;
 import org.broadinstitute.dsm.db.dao.user.UserDao;
 import org.broadinstitute.dsm.db.dto.ddp.instance.DDPInstanceDto;
 import org.broadinstitute.dsm.db.dto.ddp.participant.ParticipantDataDto;
-import org.broadinstitute.dsm.model.familymember.AddFamilyMember;
-import org.broadinstitute.dsm.model.familymember.AddFamilyMemberTest;
 import org.broadinstitute.dsm.model.participant.data.AddFamilyMemberPayload;
 import org.broadinstitute.dsm.model.participant.data.FamilyMemberDetails;
 import org.broadinstitute.dsm.model.participant.data.ParticipantData;
